Print error messages

+++ modbot/modbot.py
@@ -32,15 +32,24 @@ def add(config_file, filename, target=None):
     filename = os.path.join(path.replace(home, '~'), filename)
     target = os.path.join(target_path.replace(dotfiles, ''), target)
     config_file = os.path.join(config_path, config_file)
+
+    if filename in config[i]['link']:
+        log.error('File already in config')
+        exit(1)
+
     config[i]['link'][filename] = target
 
     with open(config_file, 'w') as f:
         yaml.safe_dump(config, f, default_flow_style=False)
 
-    log.info('Moving {0} from {1} to {2}'.format(filename, path, target_path))
-    os.rename(filename, target)
-    sys.argv[1:] = ['--config-file', config_file]
-    dotbot.main()
+    if not os.path.isfile(target):
+        log.info('Moving {0} from {1} to {2}'.format(filename, path, target_path))
+        os.rename(filename, target)
+        sys.argv[1:] = ['--config-file', config_file]
+        dotbot.main()
+    else:
+        log.error('File already linked')
+        exit(1)
 
 
 def _read_config(config_file):
